feat(object): add implementation of cache, address, and regs doms
The abstract domain is based on memory abstraction which models
memory architecture with cache. Each abstract object contains a
cache that precisely materializes memory load / store. The cache
only records the most recently used by load/store instructions.
The current cache size is one so that any loads / stores for
non mru object require to commit and update cache subdomain.

In the implementation, we also includes an alias analysis that abstracts
references alias properties. That is, we only know two references are
alias according to the DSA analysis. Such info are stored in an EUF domain.