fix(object): convert term to term op for comparison
We use UF domain to represent base addresses.
The caches in given two states refer the same MRU object if the base addresses of caches refer the same symbol in the uf domain.
The symbol is actually a term operator in crab. The comparison is based
on the value of operator but not the term.