refactor(union-find): improved performance
Implemented copy-on-write optimization. Each equivalence class keeps
now a pointer to an abstract value.  So abstract values are not copied
unless they are really needed (because they will be modified). This
has a big impact because equivalence classes are stored in maps so
they were being copied too many times.