feat(obj): improve domain reduction
(1) Add a copy-on-write method to apply reduction when the state needs to;
This will pass to join / meet operation before pair-wise join / meet on
odi map and base domain. Now, the reduction operations are performed before
subdomain operations.
(2) Add a special handler for load / store when the properties are constant.
In this case, no need to wait reduction step for overall improving performance.